Added a check to see if the Indian NSE is open and if it is, drop the latest bar of each stock
While the stock exchange is open, the latest bar of each stock is likely
shorter than the specified granularity.  This is because yfinance
attempts to return all bars up to the present in the given period.  That
includes from the end of the last full bar to the current time.  Because
the download time is non-negligible, the duration of this shortened bar
varies and affects the analysis.

This patch checks if the current time in IST (UTC+5:30) is between
9:15 and 15:30 which correlates to the NSE exchange trading hours.  If it
is open, the latest bar is dropped.